Refactor: MainViewModel 정리
- 1보 전진을 위한 3보 후퇴, 걸음마의 시작
