+++ core/common/build.gradle.kts
@@ -6,14 +6,3 @@ plugins {
 }
 
 setNamespace("core.common")
-
-dependencies {
-    //coil
-    implementation(libs.coil.compose)
-    implementation(libs.coil.network.okhttp)
-
-    implementation(libs.androidx.paging.compose)
-
-    implementation(projects.core.designsystem)
-    implementation(projects.core.model)
-}
