--- core/common/src/main/kotlin/com/kolown/porring/core/common/component/NetworkStateManager.kt
@@ -1,19 +1,0 @@
-package com.kolown.porring.core.common.component
-
-import android.content.Context
-import android.net.ConnectivityManager
-import android.net.NetworkCapabilities
-
-object NetworkStateManager {
-    fun checkNetworkState(context: Context): Boolean {
-        val connectivityManager: ConnectivityManager =
-            context.getSystemService(ConnectivityManager::class.java)
-        val network = connectivityManager.activeNetwork ?: return false
-        val actNetwork: NetworkCapabilities =
-            connectivityManager.getNetworkCapabilities(network) ?: return false
-
-        val cellularState = actNetwork.hasTransport(NetworkCapabilities.TRANSPORT_CELLULAR)
-        val wifiState = actNetwork.hasTransport(NetworkCapabilities.TRANSPORT_WIFI)
-        return cellularState || wifiState
-    }
-}
