+++ core/ui/src/main/java/com/kolown/porring/core/ui/component/DetailTopAppBar.kt
@@ -1,4 +1,4 @@
-package com.kolown.porring.core.common.component
+package com.kolown.porring.core.ui.component
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.layout.Arrangement
@@ -18,7 +18,7 @@ import androidx.compose.ui.graphics.vector.ImageVector
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.res.vectorResource
 import androidx.compose.ui.unit.dp
-import com.kolown.porring.core.common.R
+import com.kolown.porring.core.ui.R
 
 @Composable
 fun DetailTopAppBar(
