+++ core/ui/src/main/java/com/kolown/porring/core/ui/component/ShimmerEffect.kt
@@ -1,4 +1,4 @@
-package com.kolown.porring.core.common.component
+package com.kolown.porring.core.ui.component
 
 import androidx.compose.animation.core.animateFloat
 import androidx.compose.animation.core.infiniteRepeatable
