+++ core/ui/src/main/res/drawable/ic_add_circle.xml
