+++ core/ui/src/main/res/drawable/ic_cancel_circle.xml
