+++ core/ui/src/main/res/drawable/ic_detail_follow.xml
