+++ core/ui/src/main/res/drawable/ic_porring_symbol.xml
