+++ core/ui/src/main/res/drawable/icon_delete.xml
