+++ core/ui/src/main/res/drawable/icon_lost_image.xml
