+++ core/ui/src/main/res/drawable/img_heart.png
