+++ core/ui/src/main/res/drawable/img_love.png
