+++ core/ui/src/main/res/drawable/img_surprise.png
