+++ core/ui/src/main/res/drawable/img_thumb.png
