+++ feature/follower/src/main/java/com/kolown/porring/feature/follower/navigation/FollowerNavigation.kt
@@ -13,14 +13,12 @@ fun NavController.navigateFollower(navOptions: NavOptions) {
 }
 
 fun NavGraphBuilder.followerNavGraph(
-    isLoggedIn: Boolean,
+    padding: PaddingValues,
     navigateToLogin: () -> Unit,
     navigateToTheir: (String) -> Unit,
-    padding: PaddingValues,
 ) {
     composable<MainMenuRoute.Follower> {
         FollowerRoute(
-            isLoggedIn = isLoggedIn,
             navigateToLogin = navigateToLogin,
             padding = padding,
             navigateToTheir = navigateToTheir
